# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BillPro - Store Billing & Reports API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /audit-logs:
    get:
      operationId: getAuditLogs
      tags: [audit]
      summary: List a privacy-safe audit trail
      parameters:
        - { name: action, in: query, schema: { type: string } }
        - { name: entityType, in: query, schema: { type: string } }
        - { name: storeId, in: query, schema: { type: integer } }
        - { name: dateFrom, in: query, schema: { type: string, format: date-time } }
        - { name: dateTo, in: query, schema: { type: string, format: date-time } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
      responses:
        "200":
          description: Audit events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
    post:
      operationId: recordAuditEvent
      tags: [audit]
      summary: Record a safe client event such as printing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuditEventInput"
      responses:
        "201":
          description: Event recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  recorded: { type: boolean }
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Services ────────────────────────────────────────────
...
  /customers:
    get:
      operationId: getCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer with purchase history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Sales / Invoices ────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, sku, category, price, stock, createdAt]
      properties:
        id:             { type: integer }
        name:           { type: string }
        sku:            { type: string }
        category:       { type: string }
        description:    { type: ["string", "null"] }
        price:          { type: number }
        cost:           { type: ["number", "null"] }
        stock:          { type: integer }
        unit:           { type: string }
        trackSerials:   { type: boolean }
        warrantyDays:   { type: ["integer", "null"] }
        isRefurbished:  { type: boolean }
        isBundle:       { type: boolean }
        createdAt:      { type: string }
    ProductInput:
      type: object
      required: [name, sku, category, price, stock]
      properties:
        name:           { type: string }
        sku:            { type: string }
        category:       { type: string }
        description:    { type: string }
        price:          { type: number }
        cost:           { type: number }
        stock:          { type: integer }
        unit:           { type: string }
        trackSerials:   { type: boolean }
        warrantyDays:   { type: integer }
        isRefurbished:  { type: boolean }
        isBundle:       { type: boolean }
    ProductUpdate:
      type: object
      properties:
        name:           { type: string }
        sku:            { type: string }
        category:       { type: string }
        description:    { type: string }
        price:          { type: number }
        cost:           { type: number }
        stock:          { type: integer }
        unit:           { type: string }
        trackSerials:   { type: boolean }
        warrantyDays:   { type: integer }
        isRefurbished:  { type: boolean }
        isBundle:       { type: boolean }

    # ── Service ──
...
    Customer:
      type: object
      required: [id, name, createdAt]
      properties:
        id:           { type: integer }
        name:         { type: string }
        email:        { type: ["string", "null"] }
        phone:        { type: ["string", "null"] }
        address:      { type: ["string", "null"] }
        notes:        { type: ["string", "null"] }
        totalSpent:   { type: number }
        totalOrders:  { type: integer }
        createdAt:    { type: string }
    CustomerInput:
      type: object
      required: [name]
      properties:
        name:         { type: string }
        email:        { type: string }
        phone:        { type: string }
        address:      { type: string }
        notes:        { type: string }
    CustomerUpdate:
      type: object
      properties:
        name:         { type: string }
        email:        { type: string }
        phone:        { type: string }
        address:      { type: string }
        notes:        { type: string }

    # ── Line Item ──
...
    AuditLog:
      type: object
      required: [id, action, entityType, createdAt]
      properties:
        id: { type: integer }
        employeeId: { type: ["integer", "null"] }
        employeeName: { type: ["string", "null"] }
        storeId: { type: ["integer", "null"] }
        action: { type: string }
        entityType: { type: string }
        entityId: { type: ["string", "null"] }
        details: { type: ["object", "null"], additionalProperties: true }
        createdAt: { type: string, format: date-time }
    AuditEventInput:
      type: object
      required: [action, entityType]
      additionalProperties: false
      properties:
        action: { type: string, enum: [print, payment, status_change] }
        entityType: { type: string, enum: [sale, quotation, repair, backup] }
        entityId: { type: string }
...