# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BillPro - Store Billing & Reports API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditLogDetails } from './auditLogDetails';

export interface AuditLog {
  id: number;
  /** @nullable */
  employeeId?: number | null;
  /** @nullable */
  employeeName?: string | null;
  /** @nullable */
  storeId?: number | null;
  action: string;
  entityType: string;
  /** @nullable */
  entityId?: string | null;
  /** @nullable */
  details?: AuditLogDetails;
  createdAt: Date;
}